requirements.txt
# Dependencies for NYU Scholar Search Tool

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Config
pyyaml>=6.0

# Data output
pandas>=2.0.0
openpyxl>=3.1.0  # Excel support for pandas

# PDF processing
PyPDF2>=3.0.0

# Vector DB & Embeddings
chromadb>=0.4.0
openai>=1.0.0

# Environment
python-dotenv>=1.0.0

# Web app - Legacy (POC only, being replaced by FastAPI + Node frontend)
streamlit>=1.30.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# JWT verification
PyJWT[crypto]>=2.8.0
python-multipart>=0.0.6  # For form data parsing

# Authentication (optional for launch one, needed for production)
# supabase>=2.0.0
# python-jose>=3.3.0
